Refactored floats to doubles in setAxisCurrent, adapted message width to compensate

+++ psu.h
@@ -3,6 +3,6 @@
 // The value actually sent to the device should be absolute, so if x=-0.2 the current should be set to +0.2 and the negate flag
 // should be set.
 void initConnection(const char* address, const char* port);
-void setAxisCurrent(float x, float y, float z);
+void setAxisCurrent(double x, double y, double z);
 void testConnection();
 void closeConnection();
